Parameters only saved if events exist, if csv file exists it is not recreated and refactor to parameter select page

+++ src/main/java/com.smithies/analoguedataexporter/services/AnalogueReportService.java
@@ -18,6 +18,7 @@ import javax.transaction.Transactional;
 import java.io.*;
 import java.util.Iterator;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 class AnalogueReportService implements IAnalogueReportService {
@@ -38,10 +39,57 @@ class AnalogueReportService implements IAnalogueReportService {
     private AnalogueReportParametersRepository reportRepository;
 
     @Override
+    @Transactional
+    public Integer saveParameters(Short siteId, Integer channelId, long from, long to) {
+        AnalogueReportParameters parameters = new AnalogueReportParameters(siteRepository.findOne(siteId),
+                channelRepository.findOne(channelId), from, to);
+
+        // Only save if there are some valid events for the parameters provided
+        Long totalEvents = analogueRepository.countByChannel_IdAndDateBetween(parameters.getChannel().getId(),
+                parameters.getDateFrom(), parameters.getDateTo());
+        if (totalEvents==0) {
+            throw new RuntimeException("No data found for the parameters selected");
+        }
+
+        AnalogueReportParameters saved = parametersRepository.save(parameters);
+        return saved.getId();
+    }
+
+    @Override
+    public List<RawAnalogueEventsReportParametersVO> getReports() {
+        Iterable<AnalogueReportParameters> reports = reportRepository.findAll();
+        return AnalogueReportVOFactory.generateVO(reports);
+    }
+
+    @Override
+    public RawAnalogueEventsReportParametersVO getReport(Integer id) {
+        AnalogueReportParameters report = reportRepository.findById(id).orElseThrow(() -> new RuntimeException("Could not find report with id " + id));
+        return AnalogueReportVOFactory.generateVO(report);
+    }
+
+    @Override
+    public boolean reportCSVExists(Integer id) {
+        AnalogueReportParameters parameters = parametersRepository.findById(id).orElseThrow(() -> {
+            return new RuntimeException("Could not find the report parameters with id " + id);
+        });
+
+        if (new File(parametersToFilename(parameters)).exists()) {
+            return true;
+        }
+        return false;
+    }
+
+    @Override
+    public List<Boolean> reportCSVsExist(List<Integer> ids) {
+        return ids.stream().map(id -> reportCSVExists(id)).collect(Collectors.toList());
+    }
+
+    @Override
     public void downloadCsv(Integer id, HttpServletResponse response) throws IOException {
 
         // Set response headers
         response.setHeader("Content-Type", "text/csv");
+        // TODO: Make filename here be the report name
         response.setHeader("Content-Disposition", String.format("attachment;filename=%s.csv", id));
 
         // Get response output stream
@@ -51,7 +99,10 @@ class AnalogueReportService implements IAnalogueReportService {
         CSVWriter writer = new CSVWriter(new OutputStreamWriter(outputStream));
 
         // Stream file to user
-        String filename = idToFilename(id);
+        AnalogueReportParameters parameters = parametersRepository.findById(id).orElseThrow(() -> {
+            return new RuntimeException("Could not find the report parameters with id " + id);
+        });
+        String filename = parametersToFilename(parameters);
         CSVReader reader = null;
         reader = new CSVReader(new FileReader(filename));
         Iterator<String[]> iterator = reader.iterator();
@@ -72,17 +123,20 @@ class AnalogueReportService implements IAnalogueReportService {
             return new RuntimeException("Could not find the report parameters with id " + id);
         });
 
-//        List<AnalogueEvent> events = analogueRepository.findByChannel_IdAndDateBetween(parameters.getChannel().getId(),
-//                parameters.getDateFrom(), parameters.getDateTo());
+        // If csv already exists, don't bother creating it again
+        if (new File(parametersToFilename(parameters)).exists()) {
+            // File already exists so just return
+            return;
+        }
 
-        List<AnalogueEvent> events = analogueRepository.findByChannel_IdAndDateBetween(new Integer(10244700),
-                1452781798937l, 1452781811527l);
+        List<AnalogueEvent> events = analogueRepository.findByChannel_IdAndDateBetween(parameters.getChannel().getId(),
+                parameters.getDateFrom(), parameters.getDateTo());
 
         if (events.isEmpty()) {
             throw new RuntimeException("No data found for the parameters selected");
         }
 
-        String filename = idToFilename(id);
+        String filename = parametersToFilename(parameters);
         try {
             FileWriter fileWriter = new FileWriter(filename);
             CSVWriter writer = new CSVWriter(fileWriter);
@@ -104,29 +158,8 @@ class AnalogueReportService implements IAnalogueReportService {
 
     }
 
-    @Override
-    @Transactional
-    public Integer saveParameters(Short siteId, Integer channelId, long from, long to) {
-        AnalogueReportParameters params = new AnalogueReportParameters(siteRepository.findOne(siteId),
-                channelRepository.findOne(channelId), from, to);
-        AnalogueReportParameters saved = parametersRepository.save(params);
-        return saved.getId();
-    }
-
-    @Override
-    public List<RawAnalogueEventsReportParametersVO> getReports() {
-        Iterable<AnalogueReportParameters> reports = reportRepository.findAll();
-        return AnalogueReportVOFactory.generateVO(reports);
-    }
-
-    @Override
-    public RawAnalogueEventsReportParametersVO getReport(Integer id) {
-        AnalogueReportParameters report = reportRepository.findById(id).orElseThrow(() -> new RuntimeException("Could not find report with id " + id));
-        return AnalogueReportVOFactory.generateVO(report);
-
-    }
-
-    private String idToFilename(Integer id) {
-        return "analogue_report" + id + ".csv";
+    private String parametersToFilename(AnalogueReportParameters parameters) {
+        return "analogue_report_" + parameters.getInterlocking().getId() + "_" + parameters.getChannel().getId()
+                + "_" + parameters.getDateFrom() + "-" + parameters.getDateTo() + ".csv";
     }
 }
